Relocated several sections of code.
